Fix power switch, formatting, gate tests behind cfg(test)
